refactor(frontend): align admin module boundaries
