+++ frontend/features/admin/domain/formatting.tsx
@@ -2,7 +2,7 @@ import { type ApiExampleLanguage, type AppData, type Model, type ModelRoute, typ
 import { modelCategory } from "./catalog";
 import { codexImageCapableResources, findProvider, findProviderResource, isCodexSubscriptionImageModel, modelRoutesFor, stringifyForm, stringifyValue } from "./entities";
 import { guardrailBlockedDiagnostic, languageLocale, tx } from "../i18n/runtime";
-import { preferredModelCategories } from "../shared/ui";
+import { preferredModelCategories } from "./model-categories";
 
 export function initialView(): ViewKey {
   if (typeof window === "undefined") return "overview";
