+++ frontend/features/admin/domain/labels.tsx
@@ -2,7 +2,7 @@ import { type AdminResource, type AdminUser, type AppData, type SettingsTabKey }
 import { fieldSummary, projectName, stringifyValue, teamLabel } from "./entities";
 import { routeStrategyLabel } from "./formatting";
 import { displayText, tx } from "../i18n/runtime";
-import { identityProviderTemplateLabel, normalizedIdentityProviderIconKey } from "../shell/auth";
+import { identityProviderTemplateLabel, normalizedIdentityProviderIconKey } from "../shared/auth";
 
 export function compactList(value: unknown) {
   const values = Array.isArray(value) ? value.map(stringifyValue) : splitList(stringifyValue(value));
