+++ frontend/features/admin/domain/model-categories.ts
@@ -1,0 +1,46 @@
+export const modelCategoryLabels: Record<string, string> = {
+  all: "全部",
+  codex: "OpenAI Codex",
+  openai: "OpenAI",
+  claude: "Claude",
+  deepseek: "DeepSeek",
+  gemini: "Gemini",
+  qwen: "Qwen",
+  glm: "GLM",
+  kimi: "Kimi",
+  doubao: "Doubao",
+  ernie: "ERNIE",
+  baichuan: "Baichuan",
+  minimax: "MiniMax",
+  stepfun: "StepFun",
+  wanx: "WanX",
+  paddlepaddle: "PaddlePaddle",
+  microsoft: "Microsoft",
+  llama: "Llama",
+  mistral: "Mistral",
+  grok: "Grok",
+  custom: "自定义",
+};
+
+export const preferredModelCategories = [
+  "codex",
+  "openai",
+  "claude",
+  "deepseek",
+  "gemini",
+  "qwen",
+  "glm",
+  "kimi",
+  "doubao",
+  "ernie",
+  "baichuan",
+  "minimax",
+  "stepfun",
+  "wanx",
+  "grok",
+  "paddlepaddle",
+  "microsoft",
+  "llama",
+  "mistral",
+  "custom",
+];
