+++ frontend/features/admin/i18n/admin-workflows.tsx
@@ -1,0 +1,34 @@
+export const adminWorkflowTranslations = {
+  en: {
+    "SSO": "SSO",
+    "两次输入的密码不一致": "The passwords do not match",
+    "可以先使用默认额度，之后再按调用量调整。IP 白名单留空表示不限来源。": "You can start with the default limits and adjust later. Leave IP allowlist empty to allow any source.",
+    "完整 Key 只会展示一次。关闭弹窗后只能看到前后缀，后续需要通过轮换生成新 Key。": "The full Key is shown only once. After closing the modal, only the prefix and suffix are visible; rotate to generate a new Key later.",
+    "密码至少 8 位": "The password must be at least 8 characters",
+    "当前没有可选择的启用模型。请先在模型目录和路由策略里启用模型。": "No enabled models are available. Enable models in the Model Directory and routing policy first.",
+    "提交中": "Submitting",
+    "新密码": "New Password",
+    "留空表示不限制 Key 级模型白名单；实际可调用模型仍受模型目录、路由策略和项目权限约束。": "Leave blank for no Key-level model allowlist; callable models are still constrained by the Model Directory, routing policy, and project permissions.",
+    "确认新密码": "Confirm New Password",
+    "端点、Scope 与 Claim 映射": "Endpoint, Scope, and Claim Mapping",
+    "请设置新的控制台登录密码。": "Set a new console login password.",
+    "重置密码": "Reset Password",
+    "首次登录授权": "First-login Authorization",
+  },
+  ja: {
+    "SSO": "SSO",
+    "两次输入的密码不一致": "2 つのパスワードが一致しません",
+    "可以先使用默认额度，之后再按调用量调整。IP 白名单留空表示不限来源。": "まずは既定のクォータで開始し、後から利用量に応じて調整できます。IP 許可リストを空欄にすると送信元を制限しません。",
+    "完整 Key 只会展示一次。关闭弹窗后只能看到前后缀，后续需要通过轮换生成新 Key。": "完全な Key は一度だけ表示されます。ダイアログを閉じるとプレフィックスとサフィックスのみ確認でき、以降はローテーションで新しい Key を生成する必要があります。",
+    "密码至少 8 位": "パスワードは 8 文字以上にしてください",
+    "当前没有可选择的启用模型。请先在模型目录和路由策略里启用模型。": "選択できる有効なモデルがありません。先にモデルカタログとルーティングポリシーでモデルを有効化してください。",
+    "提交中": "送信中",
+    "新密码": "新しいパスワード",
+    "留空表示不限制 Key 级模型白名单；实际可调用模型仍受模型目录、路由策略和项目权限约束。": "空欄なら Key レベルのモデル許可リストは無制限です。実際に呼び出せるモデルは、モデルカタログ、ルーティングポリシー、プロジェクト権限の制約を受けます。",
+    "确认新密码": "新しいパスワードを確認",
+    "端点、Scope 与 Claim 映射": "エンドポイント、Scope、Claim のマッピング",
+    "请设置新的控制台登录密码。": "コンソールログイン用の新しいパスワードを設定してください。",
+    "重置密码": "パスワードをリセット",
+    "首次登录授权": "初回ログイン認可",
+  },
+} satisfies Record<"en" | "ja", Record<string, string>>;
