+++ frontend/features/admin/i18n/translations.tsx
@@ -1,5 +1,6 @@
 import { enTranslations } from "./en";
 import { jaTranslations } from "./ja";
+import { adminWorkflowTranslations } from "./admin-workflows";
 import { modelGovernanceTranslations } from "./model-governance";
 import { playgroundTranslations } from "./playground";
 import { providerConnectionTranslations } from "./provider-connection";
@@ -11,6 +12,6 @@ import { usageTranslations } from "./usage";
 import syntheticDNSTranslations from "./synthetic-dns";
 
 export const translations: Record<"en" | "ja", Record<string, string>> = {
-  en: { ...enTranslations, ...routingTranslations.en, ...codexImageTranslations.en, ...scopedRoutingPolicyTranslations.en, ...modelGovernanceTranslations.en, ...providerConnectionTranslations.en, ...usageTranslations.en, ...playgroundTranslations.en, ...securityTranslations.en, ...syntheticDNSTranslations.en },
-  ja: { ...jaTranslations, ...routingTranslations.ja, ...codexImageTranslations.ja, ...scopedRoutingPolicyTranslations.ja, ...modelGovernanceTranslations.ja, ...providerConnectionTranslations.ja, ...usageTranslations.ja, ...playgroundTranslations.ja, ...securityTranslations.ja, ...syntheticDNSTranslations.ja },
+  en: { ...enTranslations, ...adminWorkflowTranslations.en, ...routingTranslations.en, ...codexImageTranslations.en, ...scopedRoutingPolicyTranslations.en, ...modelGovernanceTranslations.en, ...providerConnectionTranslations.en, ...usageTranslations.en, ...playgroundTranslations.en, ...securityTranslations.en, ...syntheticDNSTranslations.en },
+  ja: { ...jaTranslations, ...adminWorkflowTranslations.ja, ...routingTranslations.ja, ...codexImageTranslations.ja, ...scopedRoutingPolicyTranslations.ja, ...modelGovernanceTranslations.ja, ...providerConnectionTranslations.ja, ...usageTranslations.ja, ...playgroundTranslations.ja, ...securityTranslations.ja, ...syntheticDNSTranslations.ja },
 };
