+++ frontend/features/admin/shared/ui.tsx
@@ -8,7 +8,7 @@ import { findProvider, modelRoutesFor } from "../domain/entities";
 import { compactNumber, routeStrategyLabel } from "../domain/formatting";
 import { enumOptionLabel, enumValueLabel, splitList } from "../domain/labels";
 import { activeLanguage, clearCustomValidity, handleRequiredFieldInvalid, selectedModelsText, selectedOptionsText, translatedCell, tx } from "../i18n/runtime";
-import { PaginationControls, usePagination } from "../views/settings-table";
+import { PaginationControls, usePagination } from "./pagination";
 
 function HiddenSelectIndicator() {
   return null;
@@ -418,50 +418,3 @@ export function ModelRouteProviders({ model, data }: { model: Model; data: AppDa
 }
 
 export const providerTypeOptions = ["mock", "openai", "openai_codex", "openai_compatible", "azure_openai", "anthropic", "gemini", "deepseek", "qwen", "local", "kronk"];
-
-export const modelCategoryLabels: Record<string, string> = {
-  all: "全部",
-  codex: "OpenAI Codex",
-  openai: "OpenAI",
-  claude: "Claude",
-  deepseek: "DeepSeek",
-  gemini: "Gemini",
-  qwen: "Qwen",
-  glm: "GLM",
-  kimi: "Kimi",
-  doubao: "Doubao",
-  ernie: "ERNIE",
-  baichuan: "Baichuan",
-  minimax: "MiniMax",
-  stepfun: "StepFun",
-  wanx: "WanX",
-  paddlepaddle: "PaddlePaddle",
-  microsoft: "Microsoft",
-  llama: "Llama",
-  mistral: "Mistral",
-  grok: "Grok",
-  custom: "自定义",
-};
-
-export const preferredModelCategories = [
-  "codex",
-  "openai",
-  "claude",
-  "deepseek",
-  "gemini",
-  "qwen",
-  "glm",
-  "kimi",
-  "doubao",
-  "ernie",
-  "baichuan",
-  "minimax",
-  "stepfun",
-  "wanx",
-  "grok",
-  "paddlepaddle",
-  "microsoft",
-  "llama",
-  "mistral",
-  "custom",
-];
