+++ tools/source-lines.mjs
@@ -50,7 +50,7 @@ export const SKIP_DIRECTORIES = new Set([
  */
 export const FROZEN = new Map([
   ["frontend/app/styles/legacy/resources.css", 1508],
-  ["frontend/features/admin/views/provider-editor.tsx", 2009],
+  ["frontend/features/admin/views/provider-editor.tsx", 2006],
 ]);
 
 export function countLines(contents) {
